Renamed package damagesources to damagesource
