+++ src/main/java/com/turtledove/withernauts/common/damagesource/PhysicalArteDamageSource.java
@@ -1,4 +1,4 @@
-package com.turtledove.withernauts.common.damagesources;
+package com.turtledove.withernauts.common.damagesource;
 
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.util.DamageSource;
@@ -8,17 +8,17 @@ import net.minecraft.util.text.TextComponentTranslation;
 
 import java.util.Random;
 
-public class ArteDamageSource extends DamageSource {
+public class PhysicalArteDamageSource extends DamageSource {
 
-    public ArteDamageSource() {
-        super("arte");
+    public PhysicalArteDamageSource() {
+        super("physical_artes");
 
         this.setDamageBypassesArmor();
     }
 
     @Override
     public ITextComponent getDeathMessage(EntityLivingBase entityLivingBaseIn) {
-        String translationKey = "death.attack.arte.player_" + new Random().nextInt(2);
+        String translationKey = "death.attack.physical_artes.player_" + new Random().nextInt(2);
 
         return new TextComponentString(entityLivingBaseIn.getDisplayName().getFormattedText() + " ")
                 .appendSibling(new TextComponentTranslation(translationKey, entityLivingBaseIn.getDisplayName()));
