+++ src/main/java/com/turtledove/withernauts/server/events/NecropolisCapabilitiesHandler.java
@@ -3,9 +3,9 @@ package com.turtledove.withernauts.server.events;
 import com.google.common.base.MoreObjects;
 import com.turtledove.withernauts.Withernauts;
 import com.turtledove.withernauts.client.gui.INecropolisItemHandler;
-import com.turtledove.withernauts.common.damagesources.ArteDamageSource;
-import com.turtledove.withernauts.common.damagesources.MeleeDamageSource;
-import com.turtledove.withernauts.common.damagesources.PhysicalArteDamageSource;
+import com.turtledove.withernauts.common.damagesource.ArteDamageSource;
+import com.turtledove.withernauts.common.damagesource.MeleeDamageSource;
+import com.turtledove.withernauts.common.damagesource.PhysicalArteDamageSource;
 import com.turtledove.withernauts.common.food.FoodEffects;
 import com.turtledove.withernauts.server.blocks.BlockHandler;
 import com.turtledove.withernauts.server.core.NecropolisPlayerData;
